single responsibility principle - new lib to generate analysis prompt
